requirements.txt
requests
dnspython
